add in/out take

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/groups/TandemMotors.java
@@ -8,7 +8,7 @@ public class TandemMotors {
         this.motors = motors;
     }
 
-    public void setPower(float power) {
+    public void setPower(double power) {
         for (DcMotor m : motors) m.setPower(power);
     }
 }
